Add validation flow
The environment that the submariner pipeline executes on prepared by
other squads.
As a result, some failures could exists already.
That could lead to a failure of the job that is not happened during the
submariner deploy or test flow.
The prerequisites check will validate the environment and will skip the
job in case prerequisites will not met.

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>