Name E2E results file with cluster names (#50)
The test flow could include multiple managed clusters.
When e2e test is running, we need to be able to identify the results
file with the clusters that it was executed on.
Set the name of the junit file with the following name staructure:

"submariner_e2e_<first_cluster_name>_<second_cluster_name>.xml