Fix internal registry info gathering (#73)
Internal registry is used during downstream deployment of Submariner.
Right before pushing the images to the internal registry, the flow takes
the internal registry url.
The url is similar in all ocp clusters:
image-registry.openshift-image-registry.svc:5000

And curently the link was taken from the hub.
But when the hub deployed on vmware platform, the error is received on
gather the internal registry link.

Fix the flow to gather the link for each dedicated cluster and not from
the hub.

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>