Fix 0.11.2 fetch iib version (#75)
* Add "@post-release" condition tag support for job execution

The "@post-release" tag is used by the pipeline after the product
version released and performs deployment from the official registry
catalog.

This is used to validate that the product functioning correctly after
the release.

- Add support for the "@post-release" tag in Jenkinsfile
- Set the proper CatalogSource for submariner depends on the deployment
  method - official/downstream

* Fix 0.11.2 iib version fetch

During downstream deployment, the iib bundle fetched to deploy the
latest operator version.
Since the 0.11.2 submariner version is no longer updated, need to add a
bigger delta for the query to fetch the required version of iib.

- Multiply the query by delta from 3 to 6