Modify tests and logs naming convention (#76)
For the easy of understanding, modify the naming convention of the
clusters to include the name of the platform (cloud) that the cluster
runs on.

The naming convention will be used as well to identify the testing
scenario of the e2e test suite.
The example of the test file name:
ACM-2.5.1-Submariner-0.12.1-AWS-GCP-Globalnet_e2e_junit.xml