Add Polarion reporting (#78)
Add reporting to Polarion flow.
The flow will create a testcases and testrun based on the junit files of
the e2e test results.

The flow requires variables that will be used for connection to polarion
and initial configurations.
The variables could be provided as evironment variables or passed as an
input file.