Fix "wait for MachineConfig" flow
When MachineConfig configuration applied, ocp master and worker nodes
are rebooted.
When the deployment is made on the SNO cluster and the node is rebooted,
the api of the cluster is not reachable.

Add "until" loop with number of iterations and sleep to ensure the
cluster api becomes reachable again.