Modify package manifest verify
The package manifest verify step ensure that the required submariner
version exists and could be deployed.
But in case some security CVE fix got released, the version of
submariner from "0.12.1" could be changed to "0.12.1+0.1656541075.p".

Since "jsonpath" does not supports the use of wildcard within the
filter, modify the query to use the jq.