Fix submariner addon version fetch (#100)
In order to identify the proper version for the subctl e2e test
execution, the submariner addon version is fetched.

But if the first cluster in the list is from the pool and in Hibernating
power state, version allocation will fail.

Add the power state query to filter only the clusters in a "Running"
state.