Add "--gather-logs" flag
Add "--gather-logs" flag to be able to gather debug logs from the
environment.
Until now, logs gathering was done at the end of the deployment
execution or in case of any error.
Adding a dedicated flag allows to gather the logs at any required time.

Also, modify the "--gather-logs true/false" to "--skip-gather-logs
true/false" to be able to specify if the logs should not be gathered
explicitly during regular execution.