Run validation on imported clusters
When a non globalnet submariner deployment is made, the flow checks for
the cluster network configuration and compares to other clusters.
If conflicting ranges detected, one of the cluster excluded from the
deployment flow to prevent conflict during submariner deployment.

As ARO and ROSA could be only imported and not created by hive, an
approach of fetching clusters information should change.

Fetch networking information directly from the clusters instead of
clusterdeployment resource as imported cluster as not available there.