Validate clusters availability (#145)
When fetching clusters for submariner deployment, it's not enough to
test that the clusters have power state - "Running".
If the cluster created from a pool and set to run but not yet claimed,
it will be still showed in "Running" state.

Check for the clusters Availability from the managed clusters output in
addition to the above check.