Modify submariner version fetch for test (#149)
When downloading subctl binary for testing, we need to get the
submariner addon version.

Currently, during the version fetch the first available cluster taken
from the clusterset.
The possible problem could be, that the clusterset contains clusters
beside the clusters with submariner deployment.
As a result, addon version could not be fetched.

Modify the function to get the first cluster from the "MANAGED_CLUSTERS"
variable that contains submariner clusters instead of the first cluster
from the clusterset.

Co-authored-by: Maxim Babushkin <mbabushk@redhat.com>