Fix Cypress "submariner_addon" test failure (#223)
Submariner_addon test fails randomly, when created clusterset takes a
few seconds to appear sometimes.
If it does not appear right away, the test fails as it unable to find
the expected clusterset.

Replace the "cy.get" with "cy.contains" and add a timeout for the
clusterset expect after creation.