Fetch explicit ocp image / acm version from quay
When fetching a version of the ocp for the managed cluster or an acm
version from quay, could be that the required version exists not on
the first page.
As a result, required version will not be fetched and deployment will
fail..
To solve it, add the following filter to the query:
"?filter_tag_name=like:<tag_version>".
It will query explicit version of the image, for example: 4.11 and
return a list of available versions:
- 4.11.59
- 4.11.58
etc...