[OCP role] - Add parallel deployment for clusters (#383)
Deployment of ocp cluster could take up to 40 minutes.
When deploying multiple clusters, it could take a very long time.

In order to decrease the time of deployment, add parallel deployment of
clusters.
All the clusters will be deployed at the same time and playbook will
watch for the state of the clusters.
That way multiple clusters deployment will take the time of a single
cluster deployment.

When deploying ocp clusters will different versions, the flow will
download "openshift-install" binary for each ocp version and append the
relevant version to the binary name.

That way, it will be possible to deploy multiple clusters with different
versions in parallel.