Make cluster passwork fetch optional
As a prerequisites of submariner deploy or test, a few steps needs to be
done.
One of the steps, is to fetch the "kubeconfig" file and kubeadmin
password value.
Some of the clusters might be missing the cluster kubeadmin password.
Make the fetch of the password optional.
Only kubeconfig file will be mandatory.