Update ImageContentSourcePolicy manifest
Recently, Openshift ROSA SME team provibited override of
registry.redhat.io explicitly.
But it's allowed to override specific namespaces.

In order to allow downstream deployment of submariner on ROSA clusters
and as a general deployment flow change, set the override mirror for
regsitry.redhat.io to the "rhacm2" namespace.

When the image will try to fetch the image from
registry.redhat.io/rhacm2, it will be overrided by the brew internal
registry.