Modify subctl fetch from official registry (#563)
When "subctl" binary was not built downstream, during the testing
pipeline, it was fetched from the upstream (github).
Once the "subctl" binary started to built downstream and shipped to the
customers, the downstream testing pipeline, used it to test and validate
the binary alongside with all other submarienr components.
But "post-publishing" flow, which should test the submariner once it
gets GA, still fetched the "subctl" from the upstream (github).
The "post-publishing" pipeline, should test all submariner components
once released to ensure the customer gets all the components in a
working state.

Modify the "post-publishing" flow in order to fetch the "subctl" binary
from the official RH registry - registry.redhat.io.
Once executed with "--downstream false" flag, the execution will expect
to get "RH_REG_USR" and "RH_REG_PSW" environment variables
to be able to fetch the required binary.

Co-authored-by: Maxim Babushkin <mbabushk@redhat.com>