Revert Jenkinsfile change for ManagedOpenshift stage
Revert the change that was made for managed openshift stage in
Jenkinsfile.
The change was made as part of creation of a job for perf / scale team.

This change created a bug, which prevented the execution of the
requested staged in CI jobs.
Reverting the change to unblock the job execution.