Revert Jenkinsfile change for ManagedOpenshift stage (#758)
Revert the change that was made for managed openshift stage in
Jenkinsfile.
The change was made as part of creation of a job for perf / scale team.

This change created a bug, which prevented the execution of the
requested staged in CI jobs.
Reverting the change to unblock the job execution.

Co-authored-by: Maxim Babushkin <mbabushk@redhat.com>